fraud-detection-trainer/pyproject.toml
[project]
name = "fraud-detection-trainer"
version = "0.1.0"
description = "Package for training a frauld detection module"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "category-encoders>=2.6.4",
    "click>=8.1.7",
    "geopy>=2.4.1",
    "loguru>=0.7.2",
    "mlflow>=2.17.0",
    "pydantic-settings>=2.6.0",
    "pydantic>=2.9.2",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.5.2",
]

[project.scripts]
train = "fraud_detection_trainer.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 90
indent-width = 4

[tool.uv]
dev-dependencies = [
    "ruff>=0.7.0",
    "ydata-profiling[notebook]>=4.11.0",
    "pandas-stubs>=2.2.3.241009",
    "ipykernel>=6.29.5",
    "pytest>=8.3.3",
    "mypy>=1.12.0",
    "seaborn>=0.13.2",
    "pdoc>=15.0.0",
]
